Edit, Delete, Location Features
Added functionality to edit and delete the ads based on preference.

Added the Map search functionality on the sidebar.

Added the Filter price on the side bar.

Added a Location Map on individual ad page which shows the location of the ad selected.

Fixed some bugs regarding the layout and map handling.